enums renamed
